move main_screen.dart into a separate folder
