+++ flutter/lib/ui/run/result_screen.dart
@@ -18,7 +18,7 @@ import 'package:mlperfbench/ui/page_constraints.dart';
 import 'package:mlperfbench/ui/run/app_bar.dart';
 import 'package:mlperfbench/ui/run/list_of_benchmark_items.dart';
 import 'package:mlperfbench/ui/run/result_circle.dart';
-import '../root/main_screen.dart';
+import '../root/main_screen/main_screen.dart';
 import 'progress_screen.dart';
 
 enum _ScreenMode { performance, accuracy }
